refactored reference editor
